show: Add damages logging

+++ compz/canvas.h
@@ -121,7 +121,7 @@ struct nemocanvas {
 	uint64_t touchid0;
 	
 	uint32_t frame_count;
-	uint32_t frame_damage;
+	uint64_t frame_damage;
 };
 
 struct nemocanvas_callback {
